Re-enable Signature entropy consumption CTS test.
Android Keystore Signature CTS test which checks that the Signature
implementation uses the SecureRandom with which it was initialized was
disabled in Android M because the Android M libcore's Signature
implementation contained a bug due to which the SecureRandom wasn't
available. This bug is now fixed and this change thus re-enables the
CTS test.

Bug: 22485587
Change-Id: I5c1f37f3b890623b618cbcee39882f59fd0f8e33