Fix CTS for DPM.resetPassword().
The issue is that DAs are no longer allowed to clear the password.
Because the test originally only had DAs, resetPassword("") in setUp()
was broken.

Now we register DO too so DPM will happily accept resetPassword("").

The resetPassowrd() API doesn't take "who" unlike other APIs --
it'll just allow it if the caller APK has DO.

Bug 25740005

Change-Id: Ie9139bab28090b8e1d85a361b92dd42ef6b1813f