Fix the CTS tests for Resources#updateConfiguration()
The previous tests used the existing Configuration of the Resources
to update the Resources, which could result in weird side effects
which were not the intention of the tests. We now create new
temporary Configurations and Resources.

Also:

* Add a new CTS test for Configuration#setTo() to make sure null
locales are copied.

* Add a new CTS test for Resources#updateConfiguration() for setting
the locales to null after setting them to another value which would
have failed previously due to a bug in Configuration#setTo() and a
bug in Resources#updateConfiguration().

Bug: 25874762
Change-Id: I7bebf6f1a443a125f98e78ce3a1f0753abce20b7