Remove BitmapFactoryTests to ARGB_4444
ARGB_4444 is deprecated.  The developer documentation states that
"as of KITKAT, any bitmap created with this configuration will be
created using ARGB_8888 instead".

In this CL, I am just removing tests that expect decodes to
ARGB_4444.  In a follow-up CL, I will test that we decode to
ARGB_8888 when ARGB_4444 is requested

Change-Id: I1bb971716954b5fd875d4d37abb4dc2229108b1f