Test for android.text.format.DateFormat.getDateFormatOrder()
There are reports of getDateFormatOrder() failing due to
unexpected pattern characters. This test iterates through
all known locales and checks the method succeeds and produces
reasonable results. For two locales ("haw" and "haw_US")
it does not because the short format is empty: bug 26397197
has been created to investigate that issue.

Bug: 26397197
Bug: https://code.google.com/p/android/issues/detail?id=82144
Change-Id: I13219a95c62101c44811827b7811b47ffcbc832f