Fix a synchronization bug in Telecom RemoteConnectionTest CTS
In rare cases, a callback set up by
setupRemoteConnectionVideoCallbacksTest
could fail to be processed by the message handler in VideoProvider
before the test case invokes a method that calls through a list of
callbacks. In these cases, the test case would fail because its expected
method would never get called due to the lack of the callback.

The fix is to insert a message into the same message handler and wait
for it to be processed before making the call from the test case.

Bug: 25167342
Change-Id: Ic41e0a445c40402f284a75899f102c793b7a8a08