Change mockito-target to mockito-api
Since mockito-target is pulled in as a static library, it contains the
entirety of junit4-target which contains
org.junit.internal.builders.AnnotatedBuilder. There appears to be a
difference between the installed version and the one that is linked in
causing IncompatibleClassChangeError when a test in the
android.core.tests.libcore.package.libcore CTS package. Switching that
to mockito-api doesn't actually link in the AnnotatedBuilder class which
is fine because it's satisfied by android.core.tests.runner which is
actually running the tests.

Bug: 26458139

(cherry picked from commit 10ecc34d18d57504a02c8fc5a060b4428b65ac6b)

Change-Id: I64bc7a79153696b3d5fed18c3eb5542b2b9d4f3f