Improve IcuTestRunner's handling of TestGroups and minor improvements
Makes sure that the IcuTestUtils handles TestFmwk.TestGroup
instances that were created with a default package properly.

Add a value for the REPORT_KEY_IDENTIFIER key when sending back
a count of the number of tests that will be run.

Cleaned up the IcuTestRunner code a little; removed duplication
of constants found elsewhere, improve documentation of existing
constants.

Change-Id: I2d027d9f9305576ad8bc9a0386be740d3ee95def