ICU: Run individual tests on all TestFmwk derived test classes
Some classes overrode TestFmwk.getTargets(String) but ignored
the supplied String (the name of the test to run) and so would
always return all tests. Previously, the code detected those
classes and treated them as if they were just one test that
included all the individual tests. There were a number of
problems with that approach:
1) If one of the tests failed then the whole class was deemed
to have failed.
2) Filtering, e.g. to hide a known error, could only work at
the level of the whole class.
3) It did not detect when there were no tests run, e.g. when it
could not find the data from which the tests were created.

This change runs each test (Target) individually. It simplifies
reporting as it has access to information on each test, rather
than having an aggregate and having to parse additional
information from the log.

It did cause some failures to occur at different times and in
different ways than previously which exposed some issues with
the error handling which necessitated some changes to
ErrorReportingRunner. As part of those changes I reverted some
formatting changes that differed from the original JUnit version
to reduce the differences and then wrapped the changed in
android-changed/end android-changed marker comments. Similar to
how the class
android.support.test.internal.runner.junit3.JUnit38ClassRunner
does it.

Switching to from run-everything to individual tests had some
impact on the icu-known-failures.txt and test-list.txt files.
In addition to those changes, the icu-known-failures.txt file
was updated with instructions on how to test individual tests.

Bug: 22023363
Change-Id: I8dbc43d28205ffe85454b05e7a30248715ab0277