New fragment tests
Add a new fragment cts module for holding tests for the Fragment APIs
in the platform. This currently includes basic lifecycle management
tests for headless fragments and fragments with views, adding,
removing, replacing, detaching and reattaching. Test that views are
created and recreated when expected. Test that getActivity returns
non-null in all expected circumstances during the lifecycle.

Change-Id: I352bcfb1afa4ab933470efc1bedc0c99902c07ff