Fix flaky Activity Transition tests.
Bug 27747253

The tests were using a static variable for holding values. It is better
to avoid that, so they were changed to use an alternative mechanism.

The return transition visibility was also being tested when it
shouldn't. When there is a finish() immediately following
finishAfterTransition, there is no guarantee that the return transition
will complete and thus set the visibility value. This value should
not be checked when that case is tested.

Change-Id: I5b3d558c913d4dae1faa1cb57f204a0be02ff33d