Add unit test for PrinterCapabilityInfo.
There are two weird cases: Theh default color mode and duplex mode do
not have to be in the allowed modes because of missing checks in the
PrinterCapabilitiesInfo constructor. As this behavior is now enshrined
in the API, we have to keep supporting this. Hence I added specific
tests for this.

Change-Id: Ic07d97f892b340d86b5f1883262a1c1e8eb0b5f4