Reduce raciness of accessibility service cts tests
The tests were relying on AccessibilityService#onDestroy to
remove references to actively running test services. That meant
that disconnected, but not destroyed, services were indistinguishable
from active, running services. If onDestroy wasn't called before
trying to set up another service, the test would fail because the
service is was trying to start appears already to be running.

Bug: 28621277
Change-Id: I3afa535203ca66efb58995a2d98988b9e5f00ac5