Add frame-rate requirement to MediaFormat used to find matched encoder
Root Cause:
The MediaFormat used to find matched test encoder and to create the test
encoder are different, which leads to the Google h.264 encoder selected
in isConcurrentEncodingDecodingSupported(), but fail to be created in
doEncoding() function due to being unable to meet the required frame
rate. If unable to find an encoder to create, the doEncoding() will
throw exception.

Solution:
Add the frame rate condition to the
isConcurrentEncodingDecodingSupported() function to screen out the
unqualified encoder at first place.

Change-Id: I2f64f744534238637af071922129c0f5684f3bd9
Signed-off-by: Xinghua Yang <xinghua.yang@mediatek.com>