Make SearchEventTest more reliable
In the emulator teams efforts to get the emulator to reliably pass CTS
tests, we've noticed that (the single test in)
android.view.cts.SearchEventTest has been flaky.  And running it
isolation on a linux workstation shows a mixture of passes and failures.
Failures seem to occur when the Activity that the test starts does not
yet have focus then the test (virtually) "presses" the search keycode.
Fix is to have the test wait for the Activity to have focus.

Please note:
* I don't work in this area, nor am I experienced in writing Java code
  for Android.  If there's a standardized way to wait for this
  condition, please advise.  Or if I'm violating any Android coding
  conventions (using System.currentTimeMillis(), for example).

* If we approve of this change (or a modified version), please advise
  what other branches I should make the same change to.  (Is there an
  nyc-cts-dev, for example?)

b/30468910

Change-Id: I3f569b514832e2257fc2c68d562e66b07696ffe1