Change the video file in MediaStore_Video_ThumbnailsTest.
The original video file format was not supported by the software H263
decoder. The new video is H264 with profile and level that are supported
by the software decoder and are sufficiently stringent according to the
requirements in the compatibility definition document.

Bug: 30232471
Change-Id: Ieb74fe85f999b73d8d1051dceb2bd54cb4ab851d
(cherry picked from commit abd1200bc242a48b8fb9c92b2a64ceffd5293f55)