DO NOT MERGE: Add test to ensure media stack has no sticky SSLv3 fallback DO NOT MERGE
The mediaserver uses the chromium HTTP stack which, in this version, had
some faulty fallback logic when combined with its OpenSSL
implementation (most other platform used the NSS implementation). It
would treat any network failure such as TCP RST as a TLS intolerant
server and the server would be stuck on the TLS intolerant list until
the mediaserver restarts which is usually only on device reset.

Bug: 31833604
Change-Id: Id603d9e9b08136eb4e6d97f5c48211ab213665a5