Redefine compatibility test not_executed counts, improve setDone
Redefine not_executed in IModuleResult and IInvocationResult. Since multiple
test runs prevent true knowledge of the quantity of not-executed tests, this
number now indicates the maximum number of not-executed tests for the module,
assuming all test runs have started. For a module to be marked "done", it must
now complete every scheduled test run without failure or missing tests. Even
if all tests are ultimately collected in one invocation, a single test run
failure mandates an additional retry to mark the module done.

bug:33211104
bug:33380715
Test: run collect-tests-only, unplug during non dEQP module
Change-Id: I8675d269389d664a7019125d48fd39cf3ab11daf