Closing test activity created by testGetRecentTasksLimitedToCurrentAPK
testGetRecentTasksLimitedToCurrentAPK() creates a SimpleActivity and
doesn’t finish it. The activity remains, and interferes with tests
run after testGetRecentTasksLimitedToCurrentAPK().

For example, in general Android,
testTimeTrackingAPI_SimpleStartExit() succeeds only by a chance,
because switching from its activity to still lingering
SimpleActivity behave similarly to switching to GEL launcher.
In ARC, the timing is different, and the launcher is different;
as a result the test doesn’t succeed by a chance, and fails.
More details in the bug:
https://b.corp.google.com/u/0/issues/29062062#comment8

The fix fixes this bug in CTS tests.
The fin will be landed in marshmallow-cts-dev, which is the AOSP
branch for cts for M release, and from there will be automerged
to ARC development branch.

Bug: 29062062
Change-Id: Ifa7eda73c5a596abd98a9537a1301098dbeefcfc

(cherry picked from commit 5db3b8167d7a588374b322b600e533f6fe24e997)