Allow failing testUsbAccessory CTS test in some cases.
(Forward porting the comment from marshmallow CTS:
http://android-review.googlesource.com/#/c/297388/.
Original patch description follows below.)

This test checks compliance with CDD for devices with USB ports(s) with
a peripheral mode support. However, Android may run on plenty of other
devices which support only the USB host mode.

A comment is added so it's clear that this test may fail in some
circumstances.

Bug: 33674098, 30750499
Test: Comment only, no changes in behavior.