CTS test for bug 33897722
Test: This is the test.

Bug 33897722 resulted in drawing with uninitialized memory. Create a
Movie with a GIF that would result in doing so without the fix, and
verify that the memory is what we expect it to be after drawing.

BUG:33897722
Change-Id: I2f320fe4f6b7deecbbb54e9f92c9fef4d26a3ded
(cherry picked from commit f1ca03c5a6268f3b44270057a518e5dd3306170e)