AlwaysOnVpnMultiStageTest: Fix error seen in newer kernels
Failures in testAlwaysOnVpnLockDown was seen specifically in a device
using 4.4 kernel.

Commit 0315e382704817b279e5693dca8ab9d89aa20b3f ("net: Fix behaviour
of unreachable, blackhole and prohibit routes") in kernel changed
the behavior of IPv4 routes to function based on the definition
of ip-route. As a result, we get error EACCESS instead of the fixed
ENETUNREACH for this scenario.

Change-Id: I77d0f48fa14bc3f0598aace63be5ce5976cb300c
(cherry picked from commit 1aae8a29a8e291eda24909dc2247d82e3af4cf0b)